refactor: reorganize directory structure by renaming lib to cli in communication and utils
refactor: update include guards from LIB to CLI for consistency

refactor: remove unused Shazam-related functions

feat: include communication module in vibra executable

refactor: relocate Shazam communication logic to separate file and modify usage in CLI
