+++ cli/CMakeLists.txt
@@ -11,10 +11,14 @@ FetchContent_Populate(args)
 add_executable(vibra
     main.cpp
     cli.cpp
+    communication/shazam.cpp
 )
 
 # Add the args include directory to the target
 target_include_directories(vibra PRIVATE ${args_SOURCE_DIR})
+
+# Add the include directory to the target
+target_include_directories(vibra PRIVATE ${CMAKE_SOURCE_DIR}/cli)
 
 # Find libcurl
 find_library(CURL_LIBRARY NAMES curl)
