+++ cli/cli.h
@@ -13,7 +13,6 @@ class CLI
     Fingerprint *getFingerprintFromMusicFile(const std::string &music_file);
     Fingerprint *getFingerprintFromStdin(int chunk_seconds, int sample_rate, int channels,
                                          int bits_per_sample, bool is_signed);
-    std::string getMetadataFromShazam(const Fingerprint *fingerprint);
 };
 
 #endif // CLI_CLI_H_
